requirements.txt
certifi==2019.3.9
chardet==3.0.4
idna==2.8
python-engineio==3.5.1
python-socketio==4.0.1
requests==2.22.0
six==1.12.0
urllib3==1.25.2
websocket-client==0.56.0
